feat: Add store check
